FFHRS-46: Allow JS execution in nextjs - Use readystate to differentiate between nextjs and drupal/storybook
